Semana7/notifications/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
flask-restful==0.3.9
gunicorn==20.1.0
PyJWT==2.1.0
requests==2.26.0
Werkzeug==2.2.2
PyDispatcher==2.0.6
pulsar-client==3.4.0
pulsar-client[avro]==3.4.0
